refactor notification system code
